requirements.txt
request
pycryptodome
pycryptodomex